requirements.txt
# Core Dependencies
crewai
crewai-tools
langchain
langchain-google-genai
langchain-community
langchain-huggingface
langchain-groq
langchain-openai
litellm
groq

# Search and Web Tools
google-search-results
duckduckgo-search
beautifulsoup4
requests
lxml

# Data Processing
pandas
numpy

# UI Framework
streamlit
streamlit-agraph

# Configuration and Environment
python-dotenv
pyyaml

# Utilities
markdown
Pillow
